+++ runtimes/eden/Cargo.toml
@@ -25,7 +25,6 @@ std = [
   "pallet-membership/std",
   "pallet-multisig/std",
   "pallet-offences/std",
-  "pallet-poa/std",
   "pallet-randomness-collective-flip/std",
   "pallet-reserve/std",
   "pallet-scheduler/std",
@@ -37,6 +36,7 @@ std = [
   "pallet-uniques/std",
   "pallet-utility/std",
   "pallet-aura/std",
+  "pallet-collator-selection/std",
   #"pallet-xcm/std",
   "polkadot-parachain/std",
   "polkadot-runtime-common/std",
@@ -88,6 +88,7 @@ runtime-benchmarks = [
   "pallet-timestamp/runtime-benchmarks",
   "pallet-uniques/runtime-benchmarks",
   "pallet-utility/runtime-benchmarks",
+  "pallet-collator-selection/runtime-benchmarks",
   #"pallet-xcm/runtime-benchmarks",
   "xcm/runtime-benchmarks",
   "xcm-builder/runtime-benchmarks",
@@ -110,7 +111,6 @@ try-runtime = [
   "pallet-membership/try-runtime",
   "pallet-multisig/try-runtime",
   "pallet-offences/try-runtime",
-  "pallet-poa/try-runtime",
   "pallet-randomness-collective-flip/try-runtime",
   "pallet-reserve/try-runtime",
   "pallet-scheduler/try-runtime",
@@ -121,6 +121,7 @@ try-runtime = [
   "pallet-uniques/try-runtime",
   "pallet-utility/try-runtime",
   "pallet-aura/try-runtime",
+  "pallet-collator-selection/try-runtime",
   "cumulus-pallet-aura-ext/try-runtime",
   "cumulus-pallet-dmp-queue/try-runtime",
   "cumulus-pallet-parachain-system/try-runtime",
@@ -199,7 +200,6 @@ frame-try-runtime = { git = "https://github.com/paritytech/substrate", default-f
 primitives = { default-features = false, path = "../../primitives" }
 pallet-allocations = { default-features = false, path = "../../pallets/allocations" }
 pallet-reserve = { default-features = false, path = "../../pallets/reserve" }
-pallet-poa = { default-features = false, path = "../../pallets/poa" }
 pallet-grants = { default-features = false, path = "../../pallets/grants" }
 pallet-mandate = { default-features = false, path = "../../pallets/mandate" }
